refactor: prevPosition 업데이트 로직 수정
